% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{maize_palette}
\alias{maize_palette}
\title{A palette generator inspired by the diversity of maize varieties}
\usage{
maize_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{color1}, \code{color2}, \code{color3},
\code{color4}, \code{color5}, \code{color6},
\code{color7}, \code{color8}}

\item{n}{Number of colors desired.
If omitted, uses all colors.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colors.
@importFrom graphics rgb rect par image text}
}
\value{
A vector of colors.
}
\description{
A palette generator inspired by the diversity of maize varieties
}
\examples{
maize_palette("color1")
maize_palette("color2")
maize_palette("color3", 5)

# If you need more colors than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colors


}
\keyword{colors}
